reverting changes related with CommandLineRunner
