--- src/main/java/com/epam/asset/tracking/config/Webconfig.java
@@ -1,46 +1,0 @@
-package com.epam.asset.tracking.config;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.BeansException;
-import org.springframework.boot.CommandLineRunner;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.ApplicationContextAware;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
-
-import java.util.Arrays;
-
-/**
- * Created on 1/29/2018.
- */
-@Configuration
-@EnableWebMvc
-public class Webconfig extends WebMvcConfigurerAdapter implements ApplicationContextAware {
-
-    Logger log = LoggerFactory.getLogger(Webconfig.class);
-
-    private ApplicationContext applicationContext;
-
-    @Bean
-    public CommandLineRunner commandLineRunner(ApplicationContext ctx) {
-        return args -> {
-
-            log.info("Let's inspect the beans provided by Spring Boot:");
-
-            String[] beanNames = ctx.getBeanDefinitionNames();
-            Arrays.sort(beanNames);
-            for (String beanName : beanNames) {
-                log.info(beanName);
-            }
-
-        };
-    }
-
-    @Override
-    public void setApplicationContext(ApplicationContext applicationContext) throws BeansException {
-        this.applicationContext = applicationContext;
-    }
-}
